#12008 [feature] 지표값 타입 Double -> Float 변경

+++ src/main/kotlin/co/brainz/itsm/sla/metricManual/dto/MetricManualDataDto.kt
@@ -12,7 +12,7 @@ import java.time.format.DateTimeFormatter
 data class MetricManualDataDto(
     val metricId: String = "",
     val referenceDate: String? = null,
-    val metricValue: Double? = null
+    val metricValue: Float? = null
 ) : Serializable {
     val formattedReferenceDate: LocalDate? = LocalDate.parse(
         referenceDate, DateTimeFormatter.ISO_DATE)
